# backend/app/main.py
from dotenv import load_dotenv
from fastapi import Depends, FastAPI, Request, Response

load_dotenv()


import os
import time
from contextlib import asynccontextmanager
from datetime import UTC, datetime

from fastapi.middleware.cors import CORSMiddleware
from fastapi.middleware.gzip import GZipMiddleware
from fastapi.responses import JSONResponse
from slowapi import _rate_limit_exceeded_handler
from slowapi.errors import RateLimitExceeded
from slowapi.middleware import SlowAPIMiddleware
from sqlalchemy import text
from sqlalchemy.orm import Session

from . import models
from .database import SessionLocal, engine
from .exceptions import AppError
from .logging_config import setup_logging
from .middleware.cache_middleware import CacheMiddleware
from .rate_limit import limiter
from .routers.v1 import v1_router
from .security import add_security_middleware
from .tasks.reminders import start_scheduler

# Setup logging with rotation


logger = setup_logging()


# NOT: Tablo oluşturma Alembic migration'larına bırakıldı.


# models.Base.metadata.create_all kaldırıldı – migration geçmişini eziyordu.


def _get_cors_origins() -> list[str]:
    """




    CORS origin listesini env'den al.




    CORS_ORIGINS doluysa onu kullan, yoksa güvenli local/prod default'lara dön.




    """

    raw = os.getenv("CORS_ORIGINS", "").strip()

    if raw:

        return [origin.strip() for origin in raw.split(",") if origin.strip()]

    return [
        "http://localhost:3000",
        "http://localhost:3001",
        "http://localhost:3002",
        "http://localhost:3003",
        "http://localhost:3004",
        "http://localhost:3005",
        "http://127.0.0.1:3000",
        "http://127.0.0.1:3001",
        "http://127.0.0.1:3002",
        "http://127.0.0.1:3003",
        "http://127.0.0.1:3004",
        "http://127.0.0.1:3005",
        "http://localhost:5173",
        "http://127.0.0.1:5173",
        "https://optiplan360.com",
        "https://www.optiplan360.com",
    ]


app = FastAPI(
    title="OPTIPLAN360",
    description="OptiPlan 360 - Üretim Yönetim Sistemi",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc",
)


# Add security middleware


add_security_middleware(app)


# Rate limiting middleware


app.state.limiter = limiter


app.add_exception_handler(RateLimitExceeded, _rate_limit_exceeded_handler)


app.add_middleware(SlowAPIMiddleware)


# Performance middleware


app.add_middleware(GZipMiddleware, minimum_size=1000)


# Cache middleware - GET isteklerini cache'le


app.add_middleware(CacheMiddleware, ttl=60)


# CORS middleware with security headers


app.add_middleware(
    CORSMiddleware,
    allow_origins=_get_cors_origins(),
    allow_credentials=True,
    allow_methods=["GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS"],
    allow_headers=["Content-Type", "Authorization", "X-Request-ID"],
    expose_headers=["X-Total-Count", "X-Response-Time"],
)


# Güvenlik header'ları security.py'deki middleware tarafından ekleniyor.


# Burada tekrar eklenmiyordu – çift kayıt kaldırıldı.


# Performance monitoring middleware


@app.middleware("http")
async def add_performance_monitoring(request, call_next):

    start_time = time.time()

    response = await call_next(request)

    process_time = time.time() - start_time

    response.headers["X-Process-Time"] = f"{process_time:.3f}s"

    if process_time > 1.0:  # Slow request warning

        logger.warning(f"Slow request: {request.method} {request.url} took {process_time:.3f}s")

    return response


# ── Global exception handler: AppError → standart JSON yanıt ──


@app.exception_handler(AppError)
async def app_error_handler(request: Request, exc: AppError) -> JSONResponse:

    return JSONResponse(
        status_code=exc.status_code,
        content={
            "error": {
                "code": exc.code,
                "message": exc.message,
                "details": [d.model_dump() for d in exc.details],
            }
        },
    )


# Include routers


app.include_router(v1_router)


# Dependency


def get_db():

    db = SessionLocal()

    try:

        yield db

    finally:

        db.close()


@app.get("/")
async def root():

    return {"message": "OptiPlan 360 API", "version": "1.0.0", "docs": "/docs", "health": "/health"}


@app.get("/auth")
async def auth_root():

    return {
        "message": "OptiPlan 360 Auth API",
        "login": "/api/v1/auth/login",
        "refresh": "/api/v1/auth/refresh",
        "me": "/api/v1/auth/me",
    }


@app.get("/orchestrator")
async def orchestrator_root():

    return {
        "message": "OptiPlan 360 Orchestrator API",
        "base": "/api/v1/orchestrator",
        "jobs": "/api/v1/orchestrator/jobs",
    }


def _seed_canonical_stations(db: Session) -> None:
    """




    MASTER_HANDOFF §0.6 – 5 kanonik üretim istasyonunu oluşturur.




    İdempotent: mevcut ise atlar.




    """

    from .models import Station

    canonical = [
        {
            "name": "Hazırlık",
            "description": "Malzeme hazırlama ve kesim hazırlığı (İki okutmalı: 1. okutma başlangıç)",
            "istasyon_durumu": "Hazır",
        },
        {
            "name": "Ebatlama",
            "description": "Parçaların boyutlandırılması (İki okutmalı: 2. okutma tamamlandı)",
            "istasyon_durumu": "Hazır",
        },
        {
            "name": "Bantlama",
            "description": "Kenar bantı uygulanması",
            "istasyon_durumu": "Hazır",
        },
        {
            "name": "Kontrol",
            "description": "Kalite kontrolü (İki okutmalı: 1. okutma teslimata hazır)",
            "istasyon_durumu": "Hazır",
        },
        {
            "name": "Teslim",
            "description": "Teslimat ve paketleme (İki okutmalı: 2. okutma teslimat yapıldı)",
            "istasyon_durumu": "Hazır",
        },
    ]

    created = 0

    for s in canonical:

        exists = db.query(Station).filter(Station.name == s["name"]).first()

        if not exists:

            db.add(
                Station(
                    name=s["name"],
                    description=s["description"],
                    active=True,
                    istasyon_durumu=s["istasyon_durumu"],
                )
            )

            created += 1

    if created:

        db.commit()

        logger.info("Kanonik istasyonlar oluşturuldu: %d adet", created)


def _seed_admin_user(db: Session) -> None:
    """




    Varsayılan admin kullanıcısını oluşturur.




    İdempotent: admin zaten varsa atlar.




    """

    from .auth import hash_password

    existing = db.query(models.User).filter(models.User.username == "admin").first()

    if existing:

        # Şifre hash'i yoksa veya boşsa yeniden oluştur

        if not existing.password_hash:

            existing.password_hash = hash_password("admin")

            db.commit()

            logger.info("Admin kullanıcısının şifresi güncellendi.")

        return

    admin_user = models.User(
        username="admin",
        email="admin@optiplan360.local",
        display_name="Admin",
        password_hash=hash_password("admin"),
        role="ADMIN",
        is_active=True,
    )

    db.add(admin_user)

    db.commit()

    logger.info("Varsayılan admin kullanıcısı oluşturuldu (admin/admin).")


def _column_exists(conn, table_name: str, column_name: str) -> bool:

    dialect = conn.dialect.name

    if dialect == "sqlite":

        rows = conn.execute(text(f"PRAGMA table_info({table_name})")).fetchall()

        return any(str(row[1]) == column_name for row in rows)

    row = conn.execute(
        text(
            """




            SELECT 1




            FROM information_schema.columns




            WHERE table_schema = current_schema()




              AND table_name = :table_name




              AND column_name = :column_name




            LIMIT 1




            """
        ),
        {"table_name": table_name, "column_name": column_name},
    ).first()

    return row is not None


def _add_column_if_missing(conn, table_name: str, column_name: str, column_sql: str) -> bool:

    if _column_exists(conn, table_name, column_name):

        return False

    conn.execute(text(f"ALTER TABLE {table_name} ADD COLUMN {column_name} {column_sql}"))

    return True


def _run_startup_schema_fixes() -> None:

    added_columns: list[str] = []

    with engine.begin() as conn:

        column_specs: list[tuple[str, str, str]] = [
            ("orders", "tracking_token", "VARCHAR(36)"),
            ("stations", "device_type", "VARCHAR"),
            ("stations", "device_model", "VARCHAR"),
            ("stations", "device_serial_number", "VARCHAR"),
            ("stations", "ip_address", "VARCHAR"),
            ("stations", "connection_type", "VARCHAR"),
            ("stations", "installation_date", "TIMESTAMP"),
            ("stations", "last_maintenance_date", "TIMESTAMP"),
            ("crm_accounts", "plaka_birim_fiyat", "DOUBLE PRECISION"),
            ("crm_accounts", "bant_metre_fiyat", "DOUBLE PRECISION"),
            ("opti_jobs", "claim_token", "VARCHAR"),
            ("opti_jobs", "xml_file_path", "VARCHAR"),
            ("opti_jobs", "result_json", "TEXT"),
            ("orders", "order_no", "INTEGER"),
            ("orders", "reminder_count", "INTEGER DEFAULT 0"),
            ("orders", "last_reminder_at", "TIMESTAMP"),
            ("optiplan_workflow_kayitlari", "teslim_tarihi", "DATE"),
            ("optiplan_workflow_kayitlari", "teslimat_adresi", "TEXT"),
            ("optiplan_workflow_kayitlari", "odeme_sekli", "VARCHAR"),
            ("optiplan_workflow_satirlari", "boy_onay", "VARCHAR"),
            ("optiplan_workflow_satirlari", "en_onay", "VARCHAR"),
            ("optiplan_workflow_satirlari", "adet_onay", "VARCHAR"),
            ("optiplan_workflow_satirlari", "boy_operator_degeri", "INTEGER"),
            ("optiplan_workflow_satirlari", "en_operator_degeri", "INTEGER"),
            ("optiplan_workflow_satirlari", "adet_operator_degeri", "INTEGER"),
            ("optiplan_workflow_satirlari", "onaylayan_id", "INTEGER"),
            ("optiplan_workflow_satirlari", "onay_zamani", "TIMESTAMP"),
            ("optiplan_workflow_satirlari", "bbox_json", "JSON"),
            ("optiplan_workflow_satirlari", "dislandi_mi", "BOOLEAN DEFAULT FALSE"),
            ("optiplan_workflow_cikarilan_satirlar", "boy_onay", "VARCHAR"),
            ("optiplan_workflow_cikarilan_satirlar", "en_onay", "VARCHAR"),
            ("optiplan_workflow_cikarilan_satirlar", "adet_onay", "VARCHAR"),
            ("optiplan_workflow_cikarilan_satirlar", "boy_operator_degeri", "INTEGER"),
            ("optiplan_workflow_cikarilan_satirlar", "en_operator_degeri", "INTEGER"),
            ("optiplan_workflow_cikarilan_satirlar", "adet_operator_degeri", "INTEGER"),
            ("optiplan_workflow_cikarilan_satirlar", "onaylayan_id", "INTEGER"),
            ("optiplan_workflow_cikarilan_satirlar", "onay_zamani", "TIMESTAMP"),
            ("optiplan_workflow_cikarilan_satirlar", "bbox_json", "JSON"),
            ("optiplan_workflow_audit_kayitlari", "user_id", "INTEGER"),
            ("optiplan_workflow_audit_kayitlari", "islem_tipi", "VARCHAR"),
        ]

        for table_name, column_name, column_sql in column_specs:

            try:

                if _add_column_if_missing(conn, table_name, column_name, column_sql):

                    added_columns.append(f"{table_name}.{column_name}")

            except Exception as exc:

                logger.warning("Schema fix atlandi (%s.%s): %s", table_name, column_name, exc)

        try:

            conn.execute(
                text(
                    "CREATE UNIQUE INDEX IF NOT EXISTS ix_orders_tracking_token ON orders(tracking_token)"
                )
            )

            conn.execute(
                text(
                    "CREATE UNIQUE INDEX IF NOT EXISTS uq_single_opti_running "
                    "ON opti_jobs(state) WHERE state = 'OPTI_RUNNING'"
                )
            )

        except Exception as exc:

            logger.warning("Schema index fix atlandi: %s", exc)

    if added_columns:

        logger.info("Schema self-heal: %d kolon eklendi.", len(added_columns))

    else:

        logger.info("Schema self-heal: eksik kolon bulunmadi.")


def _run_startup_tasks():

    # Eksik tablolari otomatik olustur (migrate edilmemislerse)

    try:

        models.Base.metadata.create_all(bind=engine, checkfirst=True)

        _run_startup_schema_fixes()

        logger.info("DB tablolari kontrol edildi / olusturuldu.")

        # Runtime UUID + order_no doldurma

        db = SessionLocal()

        try:

            import uuid

            from app.models import Order

            null_orders = db.query(Order).filter(Order.tracking_token == None).all()

            if null_orders:

                for o in null_orders:

                    o.tracking_token = str(uuid.uuid4())

                db.commit()

                logger.info(f"{len(null_orders)} eski siparise tracking_token atandi.")

            # Mevcut siparişlere sıralı order_no ata

            null_no_orders = (
                db.query(Order)
                .filter(Order.order_no == None)
                .order_by(Order.created_at.asc())
                .all()
            )

            if null_no_orders:

                from sqlalchemy import func as sa_func

                max_no = db.query(sa_func.max(Order.order_no)).scalar() or 0

                for o in null_no_orders:

                    max_no += 1

                    o.order_no = max_no

                db.commit()

                logger.info(f"{len(null_no_orders)} eski siparise order_no atandi.")

        except Exception:

            pass

        finally:

            db.close()

    except Exception as exc:

        logger.warning("Tablo olusturma hatasi (atlaniyor): %s", exc)

    start_scheduler()

    db = SessionLocal()

    try:

        _seed_canonical_stations(db)

        _seed_admin_user(db)

    except Exception as exc:

        logger.warning("Seed hatası (atlanıyor): %s", exc)

    finally:

        db.close()


@asynccontextmanager
async def lifespan(_: FastAPI):
    _run_startup_tasks()
    yield

app.router.lifespan_context = lifespan

from fastapi import WebSocket, WebSocketDisconnect

from .websockets import manager


@app.websocket("/ws")
async def websocket_endpoint(websocket: WebSocket):

    await manager.connect(websocket)

    try:

        while True:

            # Sadece bağlantıyı açık tutmak ve ping-pong için

            data = await websocket.receive_text()

    except WebSocketDisconnect:

        manager.disconnect(websocket)


@app.get("/health")
def health_check(db: Session = Depends(get_db), response: Response = None):
    """




    API sağlık kontrolü endpoint'i.




    - DB sağlıklıysa HTTP 200 + status: healthy




    - DB erişilemezse HTTP 503 + status: degraded




    Monitoring araçları (Prometheus, k8s probe, Uptime Robot) HTTP kodu okur.




    """

    db_healthy = False

    try:

        db.execute(text("SELECT 1"))

        db_healthy = True

    except Exception as e:

        logger.error(f"Health check DB hatası: {e}")

    overall = "healthy" if db_healthy else "degraded"

    if not db_healthy and response is not None:

        response.status_code = 503

    return {
        "status": overall,
        "timestamp": datetime.now(UTC).isoformat(),
        "version": "1.0.0",
        "database": "healthy" if db_healthy else "unhealthy",
        "service": "OPTIPLAN360 API",
    }


# GET /api/v1/config/permissions endpoint'i


# config_router.py'e taşındı – app/routers/config_router.py