# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OptiPlan360 API",
    "description": "Comprehensive API documentation for OptiPlan360 - Integrated Woodworking Management System",
    "version": "1.0.0",
    "contact": {
      "name": "OptiPlan360 Development Team",
      "email": "dev@optiplan360.com",
      "url": "https://optiplan360.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://api.optiplan360.com/v1",
      "description": "Production Server"
    },
    {
      "url": "https://staging-api.optiplan360.com/v1",
      "description": "Staging Server"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Login",
        "description": "Authenticate user and return access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "example": "testadmin"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "TestAdmin123!"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "token_type": {
            "type": "string",
            "example": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "example": 3600
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "testadmin"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@optiplan360.com"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "user", "operator"],
            "example": "admin"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-14T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-14T10:30:00Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "message": {
            "type": "string",
            "example": "Required field missing: customer_id"
          },
          "details": {
            "type": "object",
            "example": {
              "field": "customer_id",
              "reason": "required"
            }
          }
        }
      }
    }
  },
...
}